fix(nursing-level): 将Integer类型改为Long类型以解决主键溢出问题
由于之前的代码中，主键使用Integer类型，当数据量过大时可能会出现主键溢出的问题。为
了解决这个问题，将所有的Integer类型主键改为Long类型。此外，对于相关的Mapper和Service
接口及其实现类中的方法参数，也相应地进行了类型转换，以保持代码的一致性和稳定性。

+++ zzyl-service/src/main/java/com/zzyl/mapper/NursingPlanMapper.java
@@ -26,16 +26,16 @@ public interface NursingPlanMapper {
      * 根据id删除护理计划
      * @param id
      */
-    void deleteNursingPlan(Integer id);
-    void deleteNursingProjectPlan(Integer id);
+    void deleteNursingPlan(Long id);
+    void deleteNursingProjectPlan(Long id);
 
     /**
      * 根据id获取护理计划
      * @param
      * @return
      */
-    NursingPlan getById(Integer id);
-    List<NursingProjectPlanVo> getProjectPlan(Integer id);
+    NursingPlan getById(Long id);
+    List<NursingProjectPlanVo> getProjectPlan(Long id);
 
     /**
      *  更新护理计划
@@ -43,5 +43,5 @@ public interface NursingPlanMapper {
      */
     void updateNursingPlan(NursingPlan nursingPlan);
 
-    NursingPlan selectById(Integer id);
+    NursingPlan selectById(Long id);
 }
